% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_unique_cnvlist.R
\name{anno_cnv}
\alias{anno_cnv}
\title{Annotate overlapping of a list of unique autosomal CNVs with the following regions
\enumerate{
\item telomere / centromere
\item Gap (N-masked)
\item Segmental duplications
\item Simple Repeats
\item MHC
}}
\usage{
anno_cnv(cnv)
}
\arguments{
\item{cnv}{uniq CNV calls. Required columns: SampleID, CHROM, ID, POS, ALT, SVLEN, END, QUAL, BSR, CN}
}
\description{
Annotate overlapping of a list of unique autosomal CNVs with the following regions
\enumerate{
\item telomere / centromere
\item Gap (N-masked)
\item Segmental duplications
\item Simple Repeats
\item MHC
}
}
