% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\docType{data}
\name{centromere}
\alias{centromere}
\title{centromere}
\format{
A data frame with three variables:
\describe{
\item{chrom}{}
\item{start}{}
\item{end}{}
}
}
\usage{
centromere
}
\description{
Genomic coordinates of centromere
}
\keyword{datasets}
