% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_overlap.R
\name{computeOl_2any}
\alias{computeOl_2any}
\title{Find overlaping regions between two any data.frames of 2 lists}
\usage{
computeOl_2any(df.test, df.base)
}
\arguments{
\item{df.test}{test data.frame. Required columns:"CHROM","POS","ID","END"}

\item{df.base}{base data.frame. Required columns:"CHROM","POS","ID","END"}
}
\description{
Find overlaping regions between two any data.frames of 2 lists
}
