% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_overlap.R
\name{computeOl_2df}
\alias{computeOl_2df}
\title{Find overlaping CNVs between two data.frames of CNV lists}
\usage{
computeOl_2df(df.test, df.base)
}
\arguments{
\item{df.test}{test data.frame. Required columns:"CHROM","POS","ID","ALT","END"}

\item{df.base}{base data.frame. Required columns:"CHROM","POS","ID","ALT","END"}
}
\description{
Find overlaping CNVs between two data.frames of CNV lists
}
