% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.fix_sex}
\alias{.fix_sex}
\title{Encode sex as integer}
\usage{
.fix_sex(x)
}
\arguments{
\item{x}{vector of sex labels, encoded in any sensible way}
}
\value{
integer vector of sex codes: 0=uknown, 1=male, 2=female
}
\description{
Encode sex as integer
}
\details{
Think of the result as the count of X chromosomes.
}
