% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemix.R
\name{.prep.layout}
\alias{.prep.layout}
\title{draw the characteristic TreeMix tree with migration edges}
\usage{
.prep.layout(obj, o = NA, flip = vector(), ...)
}
\description{
draw the characteristic TreeMix tree with migration edges
}
