% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.R
\name{.sort_indiv}
\alias{.sort_indiv}
\title{Internal helper function for sorting rows in ADMIXTURE Q-matrix}
\usage{
.sort_indiv(dframe = NULL, grplab = NA, selgrp = NA, sortind = NA,
  grplabpos = NA, linepos = NA)
}
\description{
Internal helper function for sorting rows in ADMIXTURE Q-matrix
}
\details{
Code stolen whole-cloth from <https://github.com/royfrancis/pophelper/blob/master/R/pophelper.R>
}
