% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\docType{data}
\name{gap}
\alias{gap}
\title{gap}
\format{
A data frame with nine variables:
\describe{
\item{bin}{}
\item{CHROM}{}
\item{POS}{}
\item{END}{}
\item{ix}{}
\item{n}{}
\item{Size}{}
\item{Type}{}
\item{Bridge}{}
\item{ID}{}
}
}
\usage{
gap
}
\description{
Genomic coordinates of gap regions
}
\keyword{datasets}
