% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_unique_cnvlist.R
\name{gen_unique_cnvlist}
\alias{gen_unique_cnvlist}
\title{Generate a list of unique CNVs across all given samples
Re-calibrate the CNV QUAL, BSR}
\usage{
gen_unique_cnvlist(sample_cnvs)
}
\arguments{
\item{sample_cnvs}{A data.frame of Sample-level CNV calls. Required columns: SampleID, CHROM, ID, POS, ALT, SVLEN, END, QUAL, BC, CN}
}
\description{
Generate a list of unique CNVs across all given samples
Re-calibrate the CNV QUAL, BSR
}
