% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardfilter_unique_cnvlist.R
\name{hardfilter_unique_cnvlist}
\alias{hardfilter_unique_cnvlist}
\title{Apply hardfilters to annotated CNV list}
\usage{
hardfilter_unique_cnvlist(
  cnv,
  cutoff.svlen,
  cutoff.qual,
  cutoff.bsr,
  cutoff.tlmdis,
  cutoff.ctmdis,
  cutoff.olgap,
  cutoff.olsg,
  cutoff.olsrep,
  cutoff.olmhc
)
}
\arguments{
\item{cnv}{Annotated CNV data.frame Required columns: SampleID, CHROM, ID, POS, ALT, SVLEN, END, QUAL_median, BSR_median}

\item{cutoff.svlen}{cutoff of CNV size}

\item{cutoff.qual}{cutoff of CNV QUAL_median cross population}

\item{cutoff.bsr}{cutoff of CNV BSR_median cross population}

\item{cutoff.tlmdis}{cutoff of distance to telomere}

\item{cutoff.ctmdis}{cutoff of distance to centromere}

\item{cutoff.olgap}{cutoff of overlaping of CNV with gap region}

\item{cutoff.olsg}{cutoff of overlaping of CNV with segmental duplication}

\item{cutoff.olsrep}{cutoff of overlaping of CNV with simple repeats}

\item{cutoff.olmhc}{cutoff of overlaping of CNV with MHC}
}
\description{
Apply hardfilters to annotated CNV list
}
