% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{myldply}
\alias{myldply}
\title{A `dplyr`-friendly equivalent to the old `plyr::ldply()`}
\usage{
myldply(ll, fn, ..., .id = NULL)
}
\arguments{
\item{ll}{a list of objects to which to apply a function}

\item{fn}{the function to apply}

\item{...}{extra arguments passed to \code{fn}}

\item{.id}{if not \code{NULL}, name of column to use for group IDs (inferred from \code{attr(ll, "names")})}
}
\value{
a \code{tibble} with concatenated results of applying \code{fn()} to each element of \code{ll}
}
\description{
A `dplyr`-friendly equivalent to the old `plyr::ldply()`
}
\details{
Hard work of combining results is done under the hood by \code{dplyr::bind_rows()}; consult that
    function's documentation to see how incomplete rows, missing columns, conflicting data types, etc. are handled.
}

