% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_plink}
\alias{pca_plink}
\title{Perform PCA using `plink2`}
\usage{
pca_plink(bfile, skeleton = NULL, samples = NULL, max_missing = 0,
  min_maf = 0, ...)
}
\arguments{
\item{skeleton}{optional character vector fo sample IDs to use for constructing PCs}

\item{samples}{optional character vector of sample IDs to project (if using `skeleton`, should include all reference samples)}

\item{max_missing}{input filter for `plink` corresponding to `--geno`}

\item{min_maf}{input filter for `plink` corresponding to `--maf`}

\item{...}{ignored}

\item{vcf}{path to `plink` binary fileset (bed/bim/fam)}
}
\value{
An object of classes \code{tbl_df} and \code{pca_result} whose columns are the projections of
	samples onto PCs.  Eigenvalues, expressed as proportion of variance explained, are returned as an attribute.
}
\description{
Perform PCA using `plink2`
}
\details{
This function generates command-line calls to \code{plink2}, whose output is directed to a set of
	temporary files and then read into the \code{R} session. This requires the hottest version (alpha 3 or newer)
	of \code{plink2}.
}
