% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{pca_vcf}
\alias{pca_vcf}
\title{Perform PCA on a VCF file using `akt pca` and consume result}
\usage{
pca_vcf(vcf, samples = NULL, skeleton = NULL, K = 20, meta = NULL,
  region = NULL, useX = FALSE, ...)
}
\arguments{
\item{vcf}{path to VCF file; must be indexed}

\item{samples}{optional character vector of sample IDs to which the analysis should be restricted}

\item{K}{number of PCs to compute}

\item{region}{genomic region on which to perform PCA, as \code{samtools}-style region expression (eg. "chr1:1-10")}

\item{...}{ignored}
}
\value{
An object of classes \code{data.frame} and \code{pca_result} whose columns are the projections of
	samples onto PCs.  Eigenvalues, expressed as proportion of variance explained, are returned as an attribute.
}
\description{
Perform PCA on a VCF file using `akt pca` and consume result
}
\details{
This function generates a command-line call to \code{akt pca}, whose output is directed to a set of
	temporary files and then read into the \code{R} session.
}
