% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.R
\name{plot_admixture}
\alias{plot_admixture}
\title{Make standard stacked-bars plot of ancestry proportions.}
\usage{
plot_admixture(Q, label = FALSE, sort_order = NULL, pop_order = NULL, ...)
}
\arguments{
\item{Q}{a matrix of admixture proportions, melted or not}

\item{label}{show individual IDs along axis}

\item{sort_order}{apply pre-determined sort order to individuals, eg. by cluster membership}

\item{pop_order}{named vector to map arbitrary population labels ('pop1') onto meaningful ones}

\item{...}{extra arguments, ignored}
}
\description{
Make standard stacked-bars plot of ancestry proportions.
}
