% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.R
\name{plot_admixture_labelled}
\alias{plot_admixture_labelled}
\title{Make stacked-bar plot of ancestry proportions, with population labels}
\usage{
plot_admixture_labelled(
  Q,
  sort_order = NULL,
  pop_order = NULL,
  cluster_order = NULL,
  label_pops = TRUE,
  label_size = 3,
  opaque_labs = FALSE,
  label_ind = FALSE,
  nudge = 0.05,
  ymax = 1.5,
  ...
)
}
\arguments{
\item{Q}{a matrix of admixture proportions, melted or not; must have a column `pop` with population labels}

\item{sort_order}{pre-determined sort order for individuals, eg. by cluster membership}

\item{pop_order}{pre-determined sort order for populations}

\item{cluster_order}{named vector to map arbitrary ancestry-component labels ('pop1') onto meaningful ones}

\item{label_pops}{should population labels be drawn?}

\item{label_size}{text size for population labels}

\item{label_ind}{should individuals be labelled along x-axis (probaly not, if more than a few samples)}

\item{nudge}{extra buffer around population labels}

\item{ymax}{how much extra space to allow for population labels; about 1.5 is probably enough}

\item{...}{extra arguments, ignored}

\item{label}{show individual IDs along axis}
}
\description{
Make stacked-bar plot of ancestry proportions, with population labels
}
