% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POPCORN-package.R
\docType{package}
\name{POPCORN-package}
\alias{POPCORN}
\alias{POPCORN-package}
\title{POPCORN: Population-based Post-hoc Cleaning of DRAGEN CNVs R package}
\description{
A pipeline to conduct population-based post-hoc cleaning of DRAGEN CNV calls.
}
\author{
\strong{Maintainer}: Xueqing (Zoe) Zou \email{zoe.zou@astrazeneca.com}

}
\keyword{internal}
