% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_beagle_ibd}
\alias{read_beagle_ibd}
\title{Read BEAGLE-format IBD segments}
\usage{
read_beagle_ibd(ff, map = NULL, expand = FALSE, as.ranges = FALSE,
  ...)
}
\description{
Read BEAGLE-format IBD segments
}
