% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_coverage}
\alias{read_coverage}
\title{Read output from my \code{coverage.py} script}
\usage{
read_coverage(path, scrub = TRUE, as.ranges = FALSE, has.MQ0 = FALSE, ...)
}
\arguments{
\item{path}{name of file to read}

\item{scrub}{if \code{TRUE}, remove trailing suffixes (\code{*.sorted}, \code{*.merged} etc) from individual IDs}

\item{as.ranges}{if \code{TRUE}, convert to a \code{GRanges} object}
}
\value{
a dataframe
}
\description{
Read output from my \code{coverage.py} script
}
\details{
Expected input is a BED-like file with columns *chr*, *start*, *end*, *iid*, *raw read count*,
	*normalized read count* and optionally *count of MQ0 reads*, *proportion of total reads which are MQ0*.
}
