% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_fam}
\alias{read_fam}
\title{Read a PLINK family (*.fam) file}
\usage{
read_fam(ff, ...)
}
\arguments{
\item{ff}{path to a *.fam file}
}
\value{
dataframe of sample info
}
\description{
Read a PLINK family (*.fam) file
}
\details{
A *.fam file should have columns family ID, individual ID, mom ID, dad ID, sex, phenotype.  PLINK uses
  special value -9 to denote missing phenotype, for some reason, while 0=missing for mom and dad ID.
}
