% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_map}
\alias{read_map}
\title{Read a PLINK marker map (*.bim) file}
\usage{
read_map(ff, ...)
}
\arguments{
\item{ff}{path to a *.bim file}
}
\value{
dataframe of marker info
}
\description{
Read a PLINK marker map (*.bim) file
}
\details{
A *.bim file should have columns chromosome, marker ID, genetic position (cM), physical position,
  allele 1 (A1=alternate allele), allele 2 (A2=reference allele).
}
