% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.R
\name{read_Q_matrix}
\alias{read_Q_matrix}
\title{Read ancestry proportions (`Q-matrix`) from ADMIXTURE run}
\usage{
read_Q_matrix(infile, ...)
}
\arguments{
\item{infile}{path to matrix of ancestry proportions (*.Q file) from ADMIXTURE run}

\item{...}{extra arguments, ignored}
}
\description{
Read ancestry proportions (`Q-matrix`) from ADMIXTURE run
}
\details{
Assumes that estimation was performed from a PLINK *.bed file,
	and that the corresponding *.fam file resides in same directory as output.
	If standard errors for ancestrty proportions are availbale (*.Q_se file),
	they are read as well and stored in \code{attr(result, "se")}.
}

