% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemix.R
\name{read_treemix}
\alias{read_treemix}
\title{Read TreeMix output files}
\usage{
read_treemix(stem, ...)
}
\arguments{
\item{stem}{filename stem (\code{-o} from command-line) for TreeMix output}
}
\value{
a list with components including \code{m} (number of migration edges), \code{llik} (log-likelihood)
  and \code{tree} (population tree as \code{ape::phylo} object)
}
\description{
Read TreeMix output files
}

