% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{scale_shape_repeater}
\alias{scale_shape_repeater}
\title{Discrete scale of different point shapes that repeats}
\usage{
scale_shape_repeater(...)
}
\arguments{
\item{...}{passed through to underlying \code{ggplot2::scale_...()} functions}
}
\description{
Discrete scale of different point shapes that repeats
}

