% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\docType{data}
\name{sg}
\alias{sg}
\title{sg}
\format{
A data frame with three variables:
\describe{
\item{CHROM}{}
\item{POS}{}
\item{END}{}
\item{ID}{}
}
}
\usage{
sg
}
\description{
Genomic coordinates of segmental duplications
}
\keyword{datasets}
