% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.R
\name{sort_by_cluster}
\alias{sort_by_cluster}
\title{Sort samples by admixture components using crude heuristics}
\usage{
sort_by_cluster(Q, method = "all", ...)
}
\arguments{
\item{Q}{an ADMIXTURE Q-matrix from \code{read_Q_matrix()}}

\item{method}{clustering heuristic; currently only works with "all"}

\item{...}{extra arguments, ignored}
}
\value{
character vector of individual IDs in sorted order
}
\description{
Sort samples by admixture components using crude heuristics
}
