% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\docType{data}
\name{srep}
\alias{srep}
\title{srep}
\format{
A data frame with three variables:
\describe{
\item{CHROM}{}
\item{POS}{}
\item{END}{}
\item{ID}{}
}
}
\usage{
srep
}
\description{
Genomic coordinates of simple repeats
}
\keyword{datasets}
