% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\docType{data}
\name{telomere}
\alias{telomere}
\title{telomere}
\format{
A data frame with nine variables:
\describe{
\item{bin}{}
\item{chrom}{}
\item{chromStart}{}
\item{chromEnd}{}
\item{ix}{}
\item{n}{}
\item{size}{}
\item{type}{}
\item{bridge}{}
}
}
\usage{
telomere
}
\description{
Genomic coordinates of telomere
}
\keyword{datasets}
