% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{theme_slanty_x}
\alias{theme_slanty_x}
\title{Make ggplot2 x-axis labels slanted}
\usage{
theme_slanty_x(..., angle = 45, suppress_title = TRUE)
}
\arguments{
\item{angle}{angle at which to print labels (in degrees); default 45 degrees}

\item{suppress_title}{logical; if \code{TRUE} (default), don't show x-axis title}
}
\description{
Make ggplot2 x-axis labels slanted
}
