% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.R
\name{tidy}
\alias{tidy}
\title{Convert admixture proportions to tidy format.}
\usage{
tidy(Q, ...)
}
\arguments{
\item{Q}{a matrix of admixture proprtions}

\item{...}{extra arguments; if first one is a data frame, treat it as additional sample metadata,
with column \code{"iid"} as primary key}
}
\value{
a dataframe of admixture proportions (and standard errors if available),
  one row per component-individual pair
}
\description{
Convert admixture proportions to tidy format.
}
