% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture.R
\name{tidy.admixture_Q}
\alias{tidy.admixture_Q}
\title{Convert admixture proportions to tidy format.}
\usage{
\method{tidy}{admixture_Q}(Q, meta = NULL, ...)
}
\arguments{
\item{Q}{a matrix of admixture proprtions}

\item{meta}{additional sample metadata, with column \code{"iid"} as primary key}

\item{...}{extra arguments, ignored}
}
\value{
a dataframe of admixture proportions (and standard errors if available),
  one row per component-individual pair
}
\description{
Convert admixture proportions to tidy format.
}

